Defer building of error messages for Preconditions.check(State|NotNull|Argument) calls.
Two types of calls are considered.
Preconditions.checkState(foo.isBar(),
"The foo " + foo + " is not a bar");
Preconditions.checkState(foo.isBar(),
String.format("The foo %s is not a bar", foo));

Both get converted to:
Preconditions.checkState(foo.isBar(),
"The foo %s is not a bar", foo);

This is MUCH faster in the overwhelmingly common case where foo
actually is a bar because we don't need to build
(and garbage-collect) a new string each time.

A couple of examples will have been missed:
* In some cases, the tool wasn't able to automatically reduce the
line length to under 100 (about 60-70 examples).
These will be submitted in a separate, smaller CL which may
require some manual effort or a smarter tool.
* For simplicity, merge conflicts are resolved by overwriting
our version and using the version in the depot. This is safe
but means some files will be missed in this iteration.
I will mop these up in a later CL.

R=cpovirk,jmourits


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3581