Automated g4 rollback
*** Reason for rollback ***

Breaks tests

*** Original change description ***

Support removing properties defined using "this" as part of
RemoveUnusedPrototypeProperties. This requires handling properties
assigned expressions containing side-effects properly which was not
the case prior to this change.  I expect we got away with it because
prototype properties are generally assigned literals or simple variable
references.

As part of this change I also change the rules, so that any property
write (=, +=, ++) on a prototype or this property is removalable, and
other writes

... description truncated by g4 rollback ...


Revision created by MOE tool push_codebase.
MOE_MIGRATION=4161